Removed limit of the feed
The limit to the feed (20) does not reach the latest file. Using 40 as a limit works, but it seems that this is already close to the whole feed, so it is more future proof to not limit the feed. It'd be better if the feed could be retrieved using a platform as a parameter, but I haven't found that yet.