Update R.download regex
The product page now has a binary for Intel and one for ARM, but not Universal.
The format of the download URL changed due to this. Since this recipe was for Intel previously, the edit I made was to target the Intel binary.
There is now a preceding directory of `base` in the path and that is collected as part of the filename regex capture.
After making this edit, the recipe successfully downloads version 4.1.0 for Intel.