Update GIMP.download.recipe
We noticed that this recipe keeps downloading (and importing) GIMP although the version has not changed. What caused this is that the server that hosts GIMP, keeps giving different e-tags every download.

We fixed this by adding:

<key>CHECK_FILESIZE_ONLY</key>
<true/>

to URLDownloader processor, so the recipe will try to match the files via size, not e-tags.